requirements.txt
opencv-python~=4.10.0.84
pyyaml~=6.0.2